[toolchain][recipes] Move toolchains to recipes
This change trims a lot of dead code in toolchains now, in particular
code relating to chromium checkout, and depot_tools bootrstrapping. To
permit us trim all of, this change introduces an Xcode formula.

We are also adding an extra module test to make sure our chromium
checkout on windows uses the hermetic toolchain.
