+++ tools/cr/toolchains/.vpython3
@@ -1,0 +1,27 @@
+# Copyright (c) 2026 The Brave Authors. All rights reserved.
+# This Source Code Form is subject to the terms of the Mozilla Public
+# License, v. 2.0. If a copy of the MPL was not distributed with this file,
+# You can obtain one at https://mozilla.org/MPL/2.0/.
+
+python_version: "3.11"
+
+wheel: <
+  name: "infra/python/wheels/pyyaml-py3"
+  version: "version:6.0.1"
+>
+wheel: <
+  name: "infra/python/wheels/rich-py3"
+  version: "version:13.2.0"
+>
+wheel: <
+  name: "infra/python/wheels/markdown-it-py-py3"
+  version: "version:2.1.0"
+>
+wheel: <
+  name: "infra/python/wheels/mdurl-py3"
+  version: "version:0.1.2"
+>
+wheel: <
+  name: "infra/python/wheels/pygments-py3"
+  version: "version:2.14.0"
+>
