+++ tools/recipes/recipes/toolchains/xcode/build_xcode_toolchain.expected/mac.json
@@ -1,0 +1,61 @@
+[
+  {
+    "cmd": [
+      "git",
+      "clone",
+      "--depth",
+      "2",
+      "--filter=blob:none",
+      "--sparse",
+      "--branch",
+      "master",
+      "git@github.com:brave/brave-core.git",
+      "[WORKSPACE]/chromium/src/brave"
+    ],
+    "name": "clone brave-core (shallow, sparse)"
+  },
+  {
+    "cmd": [
+      "git",
+      "-C",
+      "[WORKSPACE]/chromium/src/brave",
+      "sparse-checkout",
+      "add",
+      "tools/cr/toolchains"
+    ],
+    "name": "sparse-checkout add"
+  },
+  {
+    "cmd": [
+      "git",
+      "clone",
+      "--depth",
+      "1",
+      "https://chromium.googlesource.com/chromium/tools/depot_tools",
+      "[WORKSPACE]/chromium/third_party/depot_tools"
+    ],
+    "name": "clone depot_tools"
+  },
+  {
+    "cmd": [
+      "gclient"
+    ],
+    "name": "verify gclient"
+  },
+  {
+    "cmd": [
+      "[WORKSPACE]/chromium/third_party/depot_tools/vpython3",
+      "[WORKSPACE]/chromium/src/brave/tools/cr/toolchains/build_xcode_toolchain.py",
+      "--out-dir",
+      "[WORKSPACE]/out",
+      "--chromium-tag",
+      "150.0.7841.1",
+      "--clear"
+    ],
+    "name": "build xcode toolchain"
+  },
+  {
+    "name": "$result",
+    "status": "SUCCESS"
+  }
+]
